- Fixed cg_str2sizet() and cg_str2ssizet() to cast the types.

+++ include/cybergarage/util/cstring.h
@@ -89,8 +89,8 @@ char *cg_longlong2str(CgInt64 value, char *buf, int bufSize);
 #define cg_strhex2ulong(value) (value ? strtoul(value, NULL, 16) : 0)
 #define cg_str2float(value) ((float)(value ? atof(value) : 0.0))
 #define cg_str2double(value) (value ? atof(value) : 0.0)
-#define cg_str2sizet(value) (value ? atol(value) : 0)
-#define cg_str2ssizet(value) (value ? atol(value) : 0)
+#define cg_str2sizet(value) ((size_t)(value ? atol(value) : 0))
+#define cg_str2ssizet(value) ((ssize_t)(value ? atol(value) : 0))
 
 #if defined(CG_USE_INT64)
 #if defined(__USE_ISOC99) || defined(HAVE_LONGLONG)
